# Apis/Secretaría de alumnos/SecretariaAlumnosFlaviiusNOTOCAR/swagger_server/swagger/swagger.yaml
...
swagger: "2.0"
info:
  description: "API para gestionar los datos de los alumnos"
  version: "1.0.0"
  title: "API para la gestión de alumnos"
host: "localhost:8081"
basePath: "/secretaria-alumnos"
...
schemes:
- "https"
paths:
...
  /matricula:
    post:
      tags:
      - "Matricula"
      summary: "Automatricula"
      description: "Permite a un alumno realizar su automatricula"
      operationId: "realizar_automatricula"
      consumes:
      - "application/json"
      parameters:
      - name: "id_usuario"
        in: "query"
        description: "id del usuario"
        required: true
        type: "string"
      - in: "body"
        name: "asignaturas"
        required: false
        schema:
          type: "array"
          items:
            $ref: "#/definitions/Asignatura_Matricula"
      responses:
        200:
          description: "operacion correcta"
        400:
          description: "id de usuario erroneo"
      x-swagger-router-controller: "swagger_server.controllers.matricula_controller"
  /cambiarmatricula:
    put:
      tags:
      - "Matricula"
      summary: "Cambios en matricula"
      description: "Permite la modificacion de los datos de la matricula"
      operationId: "cambiar_matricula"
      consumes:
      - "application/json"
      parameters:
      - name: "id_usuario"
        in: "query"
        description: "id del usuario"
        required: true
        type: "string"
      - in: "body"
        name: "datos"
        required: false
        schema:
          $ref: "#/definitions/Asignatura_Matricula"
      responses:
        200:
          description: "operacion correcta"
        400:
          description: "id de usuario erroneo"
      x-swagger-router-controller: "swagger_server.controllers.matricula_controller"
  /Notas:
    post:
      tags:
      - "Matricula"
      summary: "Pone la nota a un alumno con el id de matrícula"
      description: "Puntúa al alumno con la nota correspondiente"
      operationId: "poner_nota"
      consumes:
      - "application/json"
      parameters:
      - in: "body"
        name: "matriculaAsignatura"
        description: "objeto que consta de idMatricula,idGradoAsignatura y nota a\
          \ poner"
        required: true
        schema:
          $ref: "#/definitions/Matricula_asignatura"
      responses:
        200:
          description: "Alumno puntuado correctamente"
        400:
          description: "Ha habido un problema intentando puntuar al alumno."
      x-swagger-router-controller: "swagger_server.controllers.matricula_controller"
  /getMatriculaidAsignatura:
    get:
      tags:
      - "Matricula"
      summary: "Devuelve la lista de matrículas que estan asignadas a ese id de asignatura."
      description: "Devuelve las matriculas con ese id de gradoasignatura"
      operationId: "get_matriculas"
      consumes:
      - "application/json"
      parameters:
      - name: "idgradoAsignatura"
        in: "query"
        description: "id de la asignatura."
        required: true
        type: "integer"
      responses:
        200:
          description: "Matriculas devueltas correctamente"
        400:
          description: "Ha habido un problema intentando devolver las matriculas"
      x-swagger-router-controller: "swagger_server.controllers.matricula_controller"
definitions:
  Asignatura_Matricula:
    type: "object"
    properties:
      id_matriculacion:
        type: "string"
      numero_convocatoria:
        type: "integer"
      numero_matricula:
        type: "integer"
      nota:
        type: "number"
      aprobado:
        type: "boolean"
      id_alumno:
        type: "string"
      codigo_asignatura:
        type: "string"
...
  Matricula_asignatura:
    type: "object"
    properties:
      idMatricula:
        type: "integer"
      idGradoAsignatura:
        type: "integer"
      nota:
        type: "number"